initialize attribute currentUser0 and add a null check in tear down to avoid dependence of assert
git-svn-id: http://vcs.taotesting.com/svn/taoqual/branches/WorlflowEngineExt@1419 793250d4-a296-4af0-ab9f-c080ea916244

+++ test/ActivityExecutionServiceTestCase.php
@@ -15,6 +15,8 @@ class ActivityExecutionServiceTestCase extends wfEngineServiceTest {
 	 */
 	protected $service = null;
 	
+	private $currentUser0 = null;
+	
 	/**
 	 * initialize a test method
 	 */
@@ -45,7 +47,9 @@ class ActivityExecutionServiceTestCase extends wfEngineServiceTest {
 	}
 	
 	public function tearDown() {
-		$this->currentUser0->delete();
+		if (!is_null($this->currentUser0)){
+			$this->currentUser0->delete();
+		}
     }
 	
 	/**
